chore: refactor to use getStaticProps (#1557)
* feat: use getStaticProps for integrations landing

* chore: clean up page file

* chore: use getStaticProps on integration readme view

* chore: enable waypoint to allow page debugging

* chore: use getStaticProps for versioned readmes

* chore: clean up regular readme-view server

* fix: typo

* chore: use getStaticProps for component views

* fix: typo

* chore: remove static gen for non-latest

* chore: remove console.log

* chore: consolidate getStaticProps for readme views

* chore: cleanup

* chore: cleanup with lib folder

* docs: add comment to breadcrumb fn

* chore: fix up

* chore: clean up breadcrumbs a bit

* chore: use global revalidate setting

* fix: duplicate product name as in #1564

* fix: mixed server and client imports

* chore: disable waypoint integrations, done test
